+++ index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8" />
+        <link rel="icon" href="/src/favicon.ico" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="stylesheet" href="https://unpkg.com/higlass@1.12/dist/hglib.css">
+        <script src="https://cdn.tailwindcss.com"></script>
+        <style>
+            /* To resolve style conflicts with Tailwind */
+            .gosling-component { height: auto !important; }
+        </style>
+        <title>Gosling React Vite Example</title>
+    </head>
+
+    <body>
+        <script type="module" src="/src/index.jsx"></script>
+        <div id="root"></div>
+    </body>
+</html>
