+++ package.json
@@ -1,51 +1,43 @@
 {
-  "name": "gosling-react-template",
+  "name": "gosling-react-vite-example",
   "version": "0.1.0",
-  "homepage": "https://gosling-lang.github.io/gosling-react-template/",
-  "private": true,
+  "license": "MIT",
+  "homepage": "https://gosling-lang.github.io/gosling-react-vite-example/",
+  "scripts": {
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest",
+    "format": "npx eslint --fix --ext .js src/",
+    "deploy": "yarn build; touch build/.nojekyll; gh-pages -d build -t --git git"
+  },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "eslint": "^7.32.0",
-    "eslint-plugin-react": "^7.30.1",
     "gosling.js": "^0.9.28",
     "higlass": "^1.11.11",
     "pixi.js": "6.0.4",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "4.0.2",
-    "react-vega": "^7.5.1",
+    "react-vega": "^7.6.0",
+    "vega": "^5.25.0",
+    "vega-lite": "^5.9.0",
     "web-vitals": "^1.0.1"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "format": "npx eslint --fix --ext .js src/",
-    "eject": "react-scripts eject",
-    "deploy": "yarn build; touch build/.nojekyll; gh-pages -d build -t --git git"
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "eslint": "^8.40.0",
+    "eslint-plugin-react": "^7.32.2",
+    "gh-pages": "^3.1.0",
+    "vite": "^4.3.5",
+    "vitest": "^0.31.0"
   },
   "eslintConfig": {
     "extends": [
       "react-app",
       "react-app/jest"
     ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "gh-pages": "^3.1.0"
   }
 }
