--- public/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <meta name="theme-color" content="lightgray" />
-        <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-        <link rel="stylesheet" href="https://unpkg.com/higlass@1.11.3/dist/hglib.css">
-        <script src="https://cdn.tailwindcss.com"></script>
-        <style>
-            /* To resolve style conflicts with Tailwind */
-            .gosling-component { height: auto !important; }
-        </style>
-        <title>Gosling.js React Example</title>
-    </head>
-
-    <body>
-        <noscript>You need to enable JavaScript to run this app.</noscript>
-        <div id="root"></div>
-    </body>
-</html>
