+++ vite.config.js
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vite';
+import reactRefresh from '@vitejs/plugin-react-refresh';
+
+// https://vitejs.dev/config/
+export default defineConfig({
+    build: { target: 'esnext' },
+    optimizeDeps: {
+        include: ['gosling.js']
+    },
+    plugins: [reactRefresh()]
+});
